Removed exif parser helper static class.
